requirements.txt
snowflake-connector-python>=3.5.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
openai>=1.6.0 